device: Create virtual device for file input
Use of the file '-F <filename>' option allows a raw I/Q file to
be read and serves as virtual device. Sample rate must be matched
to the bandwdith / number of resource blocks.

 Resource Blocks   Sample Rate   Decimation
         6           1.92 Msps       16
        15           3.84 Msps        8
        25           5.76 Msps        4
        50          11.52 Msps        2
        75          15.36 Msps        2
       100          23.04 Msps        1

Frequency correction is currently not supported and must be
compensated separately and prior to file input (e.g. using GNU Radio).

Signed-off-by: Tom Tsou <tom@tsou.cc>